move the samplers to sampling.py
